+++ cpp/ql/test/library-tests/dataflow/external-models/protobuf.cpp
@@ -1,26 +1,9 @@
 
 // --- stub library headers ---
 
+#include "std_string.h"
+
 namespace std {
-	typedef unsigned long size_t;
-
-	template <class T> class allocator {
-	};
-
-	template<class charT> struct char_traits {
-	};
-
-	template<class charT, class traits = char_traits<charT>, class Allocator = allocator<charT> >
-	class basic_string {
-	public:
-		basic_string();
-		basic_string(const charT* s, const Allocator& a = Allocator());
-		const charT* data() const;
-		size_t size() const;
-	};
-
-	typedef basic_string<char> string;
-
 	class istream {
 	public:
 		istream();
