+++ cpp/ql/test/library-tests/dataflow/external-models/sinks.expected
@@ -1,5 +1,5 @@
-| asio_streams.cpp:135:29:135:39 | recv_buffer | remote-sink |
-| asio_streams.cpp:145:29:145:39 | send_buffer | remote-sink |
+| asio_streams.cpp:122:29:122:39 | recv_buffer | remote-sink |
+| asio_streams.cpp:132:29:132:39 | send_buffer | remote-sink |
 | test.cpp:12:10:12:10 | 0 | test-sink |
 | test.cpp:14:10:14:10 | x | test-sink |
 | test.cpp:18:10:18:10 | y | test-sink |
