+++ cpp/ql/test/library-tests/dataflow/external-models/sources.expected
@@ -1,4 +1,4 @@
-| asio_streams.cpp:129:34:129:44 | read_until output argument | remote |
+| asio_streams.cpp:116:34:116:44 | read_until output argument | remote |
 | azure.cpp:253:48:253:60 | call to GetBodyStream | remote |
 | azure.cpp:273:52:273:61 | call to GetHeaders | remote |
 | azure.cpp:277:38:277:44 | call to GetBody | remote |
